refactor(processors): split processors into package
