+++ src/smart_home/server/processors/__init__.py
@@ -1,0 +1,5 @@
+from smart_home.server.processors.register import RegisterProcessor
+from smart_home.server.processors.response import ResponseProcessor
+from smart_home.server.processors.state_change import StateChangeProcessor
+
+__all__ = ["RegisterProcessor", "ResponseProcessor", "StateChangeProcessor"]
