+++ src/smart_home/server/processors/response.py
@@ -1,0 +1,3 @@
+class ResponseProcessor:
+    async def handle(self, event) -> None:
+        print(f"[ResponseProcessor] Received response from device {event.device_id}")
