+++ src/smart_home/server/processors/state_change.py
@@ -1,0 +1,3 @@
+class StateChangeProcessor:
+    async def handle(self, event) -> None:
+        print(f"[StateChangeProcessor] Received state change from device {event.device_id}")
